Task5: A function that returns the natural square root of a number using recursion

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -11,7 +11,7 @@
 
 int check_for_square(int x, int y)
 {
-	int root = (x + y) / 2;
+	long unsigned int root = (x + y) / 2;
 
 	if (y == 1)
 		return (1);
